w3c-reports/12092014/gtk/runned-tests/html/semantics/grouping-content/the-ol-element/grouping-ol-expected.txt

Description

This test validates the ol element.

Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item
list item
Ordered List

list item
list item












list item

PASS The prototype for OL is HTMLOListElement.prototype 
FAIL 'reversed' property should be defined on OL. assert_own_property: expected property "reversed" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:172:32
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:170:13)
FAIL 'start' property should be defined on OL. assert_own_property: expected property "start" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:177:32
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:175:13)
FAIL 'type' property should be defined on OL. assert_own_property: expected property "type" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:182:32
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:180:13)
PASS OL's 'reversed' IDL property reflects content attribute. 
PASS OL's 'start' IDL property reflects content attribute. 
PASS OL's 'type' IDL property reflects content attribute. 
PASS IDL 'reversed' attribute value false when content attribute absent 
PASS IDL 'reversed' attribute value true when content attribute exists 
PASS Changing IDL 'reversed' property changes list's reversed property. 
PASS Default start value for non-reversed list should be 1 
PASS IDL and content attribute parse start of '.5' correctly. 
PASS IDL and content attribute parse start of 'A' correctly. 
PASS Default start value (if none provided) for reversed list = child li elements. 
PASS Default start value (if failed to parse) for reversed list = child li elements. 
PASS Default start value for reversed list = child li elements (even with tons of other child elements). 
FAIL Adding child element to reversed list adds 1 to start value assert_equals: Adding child element to reversed list adds 1 to start value expected 4 but got 3(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:267:26
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/grouping-content/the-ol-element/grouping-ol.html:264:13)
PASS Deleting child element from reversed list reduces start value by 1 
PASS IDL and content attribute parse start of '2' correctly. 
PASS IDL and content attribute parse start of '-10' correctly. 
PASS IDL and content attribute parse start of '4.03' correctly. 
PASS IDL and content attribute parse start of '-4.03' correctly. 
PASS IDL and content attribute parse start of '4.9' correctly. 
PASS IDL and content attribute parse start of '-4.9' correctly. 
PASS IDL and content attribute parse start of '7e2' correctly. 
